First Version of the My Finances Application
The application currently has basic terminal functionality. Most of the
basic logic is complete but you can't access all of it from the terminal
window yet. The database interactions are also not enabled yet. The next
version will be implementing a GUI which will create the environment
needed to flesh out the rest of the program's logic.